Move results into a different target
