--- Sources/ResultModel/Placeholder.swift
@@ -1,1 +1,0 @@
-import Foundation
