+++ Tests/ResultModelTests/ResultDataTests.swift
@@ -31,7 +31,8 @@
 //
 
 import XCTest
-@testable import JsonModel
+import JsonModel
+@testable import ResultModel
 
 class ResultDataTests: XCTestCase {
